Organised desired capabilities into a seperate file
